requirements.txt
Flask-Admin==1.0.9
Flask-Login==0.2.11
Flask-MongoEngine==0.7.1
Pillow==2.6.1
-e git://github.com/ajduncan/downspout.git@v0.0.2#egg=downspout

